pyproject.toml
[project]
name = "cubinho"
version = "0.1.0"
description = "Assistente de engenharia de software"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
]